+++ app/inc/head.php
@@ -1,3 +1,5 @@
+<?php defined('APPBASE') or die; ?>
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">