+++ app/utils/index.php
@@ -1,1 +1,5 @@
-<?php // silence is golden
+<?php defined('APPBASE') or die;
+
+/**
+ * index.php
+ */